Add support for STAMP
Also make OPEN and STAMP op ids consistent between python and C.
Comment out code for compilation without stat support.
Opportunistically fix trailing whitespace.

+++ step1/libzhpe_stats/libzhpe_stats.c
@@ -58,7 +58,7 @@ static_assert(sizeof(struct zhpe_stats_record)%64 == 0, "foo");
         : "c" (counter))
 
 /* copied from perf_event_open man page */
-static long perf_event_open(struct perf_event_attr *pea, pid_t pid,
+static int perf_event_open(struct perf_event_attr *pea, pid_t pid,
                int cpu, int group_fd, unsigned long flags)
 {
     int ret;
